FIX: Incorrect default branch name

+++ create-layer.sh
@@ -55,4 +55,11 @@ fi
 mv "$host_temp_dir"/lambda-layer.zip "${layername}".zip
 
 echo "Finishing up - find your layer file as ${layername}.zip"
-rm -rf "$host_temp_dir"
+
+if [ -z "$CLEANUP" ]; then
+  # Docker in github actions runs under a local root user, I cannot delete the temp dir after that
+  echo "Skipping cleanup"
+else
+  rm -rf "$host_temp_dir"
+fi
+
